+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as HdButton } from './Button.vue'
+export { default as HdButtonGroup } from './ButtonGroup.vue'